added more basic metrics
