+++ docs/metrics.md
@@ -1,0 +1,3 @@
+## Extension for ML Metrics/Losses
+
+::: polars_ds.complex.metrics
