+++ docs/polars_ds.md
@@ -2,4 +2,4 @@
 
 ::: polars_ds
     options:
-        filters: ["!(NumExt|StatsExt|StrExt|ComplexExt)", "^__init__$"]
+        filters: ["!(NumExt|StatsExt|StrExt|ComplexExt|MetricExt)", "^__init__$"]
