+++ mkdocs.yml
@@ -9,6 +9,7 @@ nav:
 - Numerical Extension: num.md
 - Stats Extension: stats.md
 - String Extension: str2.md
+- ML Metrics/Loss Extension: metrics.md
 - Additional Expressions: polars_ds.md
 
 theme:
